+++ db/src/main/java/ru/alekseiadamov/db/dto/ProductListParamsDTO.java
@@ -4,12 +4,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class ProductListParamsDTO extends PageParamsDTO {
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private Integer category;
     private String brand;
 }
